refactor(snapp): routing in multiple components
