+++ apps/sensenet/src/components/MainRouter.tsx
@@ -1,6 +1,6 @@
 import { LoginState } from '@sensenet/client-core'
 import React, { lazy, Suspense, useEffect, useRef } from 'react'
-import { Route, RouteComponentProps, Switch, withRouter } from 'react-router'
+import { Redirect, Route, RouteComponentProps, Switch, withRouter } from 'react-router'
 import { LoadSettingsContextProvider, RepositoryContext, useSession } from '@sensenet/hooks-react'
 import { usePersonalSettings } from '../hooks'
 import { FullScreenLoader } from './FullScreenLoader'
@@ -53,20 +53,16 @@ const MainRouter: React.StatelessComponent<RouteComponentProps> = props => {
   return (
     <ErrorBoundary FallbackComponent={ErrorBoundaryWithDialogs}>
       <Route
-        render={() => (
-          <div style={{ width: '100%', height: '100%' }}>
-            <Suspense fallback={<FullScreenLoader />}>
-              <Switch>
+        render={({ location }) =>
+          sessionContext.state === LoginState.Unauthenticated || !personalSettings.lastRepository ? (
+            <Redirect to={{ pathname: '/login', state: { from: location } }} />
+          ) : (
+            <div style={{ width: '100%', height: '100%' }}>
+              <Suspense fallback={<FullScreenLoader />}>
                 <Route
                   path="/personalSettings"
                   render={() => {
                     return <PersonalSettingsEditor />
-                  }}
-                />
-                <Route
-                  path="/login"
-                  render={() => {
-                    return <LoginComponent />
                   }}
                 />
                 <Route
@@ -81,106 +77,98 @@ const MainRouter: React.StatelessComponent<RouteComponentProps> = props => {
                     return <EventListComponent />
                   }}
                 />
+                <Switch>
+                  <Route
+                    path="/:repo/browse/:browseData?"
+                    render={routeProps => {
+                      return <ExploreComponent {...routeProps} />
+                    }}
+                  />
+                  <Route
+                    path="/:repo/search/:queryData?"
+                    render={routeProps => {
+                      return (
+                        <LoadSettingsContextProvider>
+                          <SearchComponent {...routeProps} />
+                        </LoadSettingsContextProvider>
+                      )
+                    }}
+                  />
 
-                {/** Requires login */}
-                {sessionContext.state === LoginState.Unauthenticated || !personalSettings.lastRepository ? (
-                  <LoginComponent />
-                ) : sessionContext.state === LoginState.Authenticated ? (
-                  <Switch>
-                    <Route
-                      path="/:repo/browse/:browseData?"
-                      render={routeProps => {
-                        return <ExploreComponent {...routeProps} />
-                      }}
-                    />
-                    <Route
-                      path="/:repo/search/:queryData?"
-                      render={routeProps => {
-                        return (
-                          <LoadSettingsContextProvider>
-                            <SearchComponent {...routeProps} />
-                          </LoadSettingsContextProvider>
-                        )
-                      }}
-                    />
-
-                    <Route
-                      path="/:repo/saved-queries"
-                      render={() => {
-                        return (
-                          <LoadSettingsContextProvider>
-                            <SavedQueriesComponent />
-                          </LoadSettingsContextProvider>
-                        )
-                      }}
-                    />
-                    <Route
-                      path="/:repo/setup"
-                      render={() => {
-                        return <SetupComponent />
-                      }}
-                    />
-                    <Route
-                      path="/:repo/trash"
-                      render={() => {
-                        return <TrashComponent />
-                      }}
-                    />
-                    <Route
-                      path="/:repo/editBinary/:contentId?"
-                      render={() => {
-                        return <EditBinary />
-                      }}
-                    />
-                    <Route
-                      path="/:repo/editProperties/:contentId?"
-                      render={() => {
-                        return <EditProperties />
-                      }}
-                    />
-                    <Route
-                      path="/:repo/preview/:documentId?"
-                      render={() => {
-                        return <DocumentViewerComponent previousLocation={previousLocation.current} />
-                      }}
-                    />
-                    <Route path="/:repo/wopi/:documentId/:action?">
-                      <WopiPage />
-                    </Route>
-                    <Route
-                      path="/:repo/dashboard/:dashboardName?"
-                      render={routeParams => {
-                        return (
-                          <RepositoryContext.Consumer>
-                            {repo => <DashboardComponent repository={repo} {...routeParams} />}
-                          </RepositoryContext.Consumer>
-                        )
-                      }}
-                    />
-                    <Route
-                      path="/:repo/"
-                      render={routeParams => {
-                        return (
-                          <RepositoryContext.Consumer>
-                            {repo => <DashboardComponent repository={repo} {...routeParams} />}
-                          </RepositoryContext.Consumer>
-                        )
-                      }}
-                    />
-                    <Route
-                      path="/"
-                      render={routeParams => {
-                        return <DashboardComponent {...routeParams} />
-                      }}
-                    />
-                  </Switch>
-                ) : (
-                  <FullScreenLoader />
-                )}
-              </Switch>
-            </Suspense>
-          </div>
-        )}
+                  <Route
+                    path="/:repo/saved-queries"
+                    render={() => {
+                      return (
+                        <LoadSettingsContextProvider>
+                          <SavedQueriesComponent />
+                        </LoadSettingsContextProvider>
+                      )
+                    }}
+                  />
+                  <Route
+                    path="/:repo/setup"
+                    render={() => {
+                      return <SetupComponent />
+                    }}
+                  />
+                  <Route
+                    path="/:repo/trash"
+                    render={() => {
+                      return <TrashComponent />
+                    }}
+                  />
+                  <Route
+                    path="/:repo/editBinary/:contentId?"
+                    render={() => {
+                      return <EditBinary />
+                    }}
+                  />
+                  <Route
+                    path="/:repo/editProperties/:contentId?"
+                    render={() => {
+                      return <EditProperties />
+                    }}
+                  />
+                  <Route
+                    path="/:repo/preview/:documentId?"
+                    render={() => {
+                      return <DocumentViewerComponent previousLocation={previousLocation.current} />
+                    }}
+                  />
+                  <Route path="/:repo/wopi/:documentId/:action?">
+                    <WopiPage />
+                  </Route>
+                  <Route
+                    path="/:repo/dashboard/:dashboardName?"
+                    render={routeParams => {
+                      return (
+                        <RepositoryContext.Consumer>
+                          {repo => <DashboardComponent repository={repo} {...routeParams} />}
+                        </RepositoryContext.Consumer>
+                      )
+                    }}
+                  />
+                  <Route
+                    path="/:repo/"
+                    render={routeParams => {
+                      return (
+                        <RepositoryContext.Consumer>
+                          {repo => <DashboardComponent repository={repo} {...routeParams} />}
+                        </RepositoryContext.Consumer>
+                      )
+                    }}
+                  />
+                  <Route
+                    path="/"
+                    render={routeParams => {
+                      return <DashboardComponent {...routeParams} />
+                    }}
+                  />
+                </Switch>
+              </Suspense>
+            </div>
+          )
+        }
       />
     </ErrorBoundary>
   )
