+++ apps/sensenet/src/components/sn-injector.ts
@@ -1,5 +1,5 @@
 import { Injector, VerboseConsoleLogger } from '@sensenet/client-utils'
-import { EventLogger } from './services/EventLogger'
+import { EventLogger } from '../services/EventLogger'
 
 export const snInjector = new Injector()
 snInjector.options.owner = 'SnApp'
