+++ apps/sensenet/src/services/CommandProviders/index.ts
@@ -1,0 +1,7 @@
+export * from './CheatCommandProvider'
+export * from './CustomActionCommandProvider'
+export * from './HelpCommandProvider'
+export * from './HistoryCommandProvider'
+export * from './InFolderSearchCommandProvider'
+export * from './NavigationCommandProvider'
+export * from './QueryCommandProvider'
